refactor: rename components to shared and rm old styles
