+++ client/src/App.tsx
@@ -1,5 +1,4 @@
 import { BrowserRouter } from 'react-router-dom'
-import './App.css'
 import AppRoute from './routes/AppRoute'
 function App() {
   return (
