+++ client/src/index.css
@@ -1,38 +1,17 @@
-:root {
-  font-family: 'Karla', sans-serif;
-  font-family: 'Plus Jakarta Sans', sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-text-size-adjust: 100%;
-}
+@import 'tailwindcss/base';
+@import 'tailwindcss/components';
+@import 'tailwindcss/utilities';
 
+/* Global Styles */
 body {
-  margin: 0;
-}
-ul {
-  list-style: none;
+  @apply font-open-sans;
 }
 
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-a {
-  text-decoration: none;
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  @apply font-raleway;
 }
